compatible gcc and remove all warning

+++ canonical_huffman.c
@@ -10,8 +10,8 @@ void make_canonical_huffman_codes(uint8_t *huff_size, uint16_t *huff_code, const
 		nb = bits_table[i];
 		for (j = 0; j<nb; j++) {
 			sym = val_table[k++];
-			huff_size[sym] = i + 1;
-			huff_code[sym] = code;
+			huff_size[sym] = (uint8_t)(i + 1);
+			huff_code[sym] = (uint16_t)code;
 			code++;
 		}
 		code <<= 1;
